Reduce pydantic eager load
