+++ src/zrb/session_state_logger/any_session_state_logger.py
@@ -1,19 +1,21 @@
 import datetime
+from typing import TYPE_CHECKING
 from abc import ABC, abstractmethod
 
-from zrb.session_state_log.session_state_log_model import (
-    SessionStateLog,
-    SessionStateLogList,
-)
+if TYPE_CHECKING:
+    from zrb.session_state_log.session_state_log import (
+        SessionStateLog,
+        SessionStateLogList,
+    )
 
 
 class AnySessionStateLogger(ABC):
     @abstractmethod
-    def write(self, session_log: SessionStateLog):
+    def write(self, session_log: "SessionStateLog"):
         pass
 
     @abstractmethod
-    def read(self, session_name: str) -> SessionStateLog:
+    def read(self, session_name: str) -> "SessionStateLog":
         pass
 
     @abstractmethod
@@ -24,5 +26,5 @@ class AnySessionStateLogger(ABC):
         max_start_time: datetime.datetime,
         page: int = 0,
         limit: int = 10,
-    ) -> SessionStateLogList:
+    ) -> "SessionStateLogList":
         pass
